requirements.txt
aiohttp==3.10.5
aiosmtpd==1.4.4.post2
python-dotenv==1.0.1
requests==2.32.3